ensuring proper user-access to accounts
ensuring accounts can only be accessed by their creator and not by other users

+++ lib/src/notifiers/account_notifier.dart
@@ -19,8 +19,12 @@ class AccountNotifier extends StateNotifier<List<Account>> {
 
   Future<void> fetchAccounts() async {
     try {
-      var accounts = await _firestoreService.fetchAll();
-      state = accounts.cast<Account>();
+
+      final user = ref.read(authServiceProvider).currentUser;
+      if (user == null) return;
+
+      var accounts = await _firestoreService.fetchAll(user.uid);
+      state = accounts;
     } catch (e) {
       print("Error fetching accounts: $e");
     }
@@ -28,19 +32,23 @@ class AccountNotifier extends StateNotifier<List<Account>> {
 
   Future<void> addAccount(Account account) async {
     try {
-      print("Adding account: ${account.toMap()}");
-      await _firestoreService.create(account.toMap(), account.id);
+      final user = ref.read(authServiceProvider).currentUser;
+      if (user == null) return;
+
+      final newAccount = account.copyWith(creatorId: user.uid);
+      await _firestoreService.create(newAccount.toMap(), newAccount.id, user.uid);
+
 
       // Initialize default categories
       for (var category in defaultCategories) {
         final newCategory = Category(
           id: DateTime.now().millisecondsSinceEpoch.toString(),
-          accountId: account.id,
+          accountId: newAccount.id,
           name: category['name'],
           type: category['type'],
-          creatorId: account.creatorId,
+          creatorId: newAccount.creatorId,
         );
-        await _categoryService.create(newCategory.toMap(), newCategory.id);
+        await _categoryService.create(newCategory.toMap(), newCategory.id, user.uid);
       }
 
       state = [...state, account];
@@ -51,8 +59,11 @@ class AccountNotifier extends StateNotifier<List<Account>> {
 
   Future<void> removeAccount(String accountId) async {
     try {
+      final user = ref.read(authServiceProvider).currentUser;
+      if (user == null) return;
+
       // Fetch and delete all activities related to the account
-      var activities = await _activityService.fetchAll();
+      var activities = await _activityService.fetchAll(user.uid);
       for (var activity in activities) {
         if (activity.accountId == accountId) {
           await _activityService.delete(activity.id);
